Set license to MIT
This way tool for automatic license checking are able to see that the license for this project is a standard MIT license. Otherwise, it is detected as a non-standard license (for example on https://crates.io/crates/base-x). 

Also when `license` is properly set then the special make targets like `cargo-crates-licenses` designed for easier maintenance of FreeBSD Rust ports are producing nice results.